Split Archetype search results into a separate view
