+++ decksite/templates/archetypesearch.mustache
@@ -1,0 +1,64 @@
+<section>
+    <h2>Search for Decks</h2>
+
+    <form method="post">
+        <div>
+            <label for="q">With these cards (separated by newlines)</label>
+            <textarea name="q">{{ query }}</textarea>
+        </div>
+        <div>
+            <label for="notq">Without these cards</label>
+            <textarea name="notq">{{ notquery }}</textarea>
+        </div>
+        <button type="submit">Search by Card</button>
+    </form>
+
+    <form method="post">
+        <p>Use this tool to automatically find cases where identical decks have been assigned different archetypes.</p>
+        <input type="hidden" name="find_conflicts" value="true"></input>
+        <button type="submit" name="conflicts">Search for Conflicts</button>
+    </form>
+
+    {{#has_search_results}}
+        <form method="post" class="inline">
+            <table id="search-results" data-toggle="checkboxes" data-range="true">
+                <thead>
+                    <tr>
+                        <th>Colors</th>
+                        <th>Name</th>
+                        <th>Archetype</th>
+                        <th>
+                            <a href="#search-results" data-toggle="checkboxes" data-action="check">☑</a>
+                            |
+                            <a href="#search-results" data-toggle="checkboxes" data-action="uncheck">☐</a>
+                        </th>
+                        <th>Season</th>
+                    </tr>
+                </thead>
+                <tbody>
+                    {{#search_results}}
+                        <tr>
+                            <td>{{{ colors_safe }}}</td>
+                            <td title="{{ decklist }}">
+                                <a href="{{ url }}">{{ name }}</a>
+                            </td>
+                            <td>
+                                <a href="{{ archetype_url }}">{{ archetype_name }}</a>
+                            </td>
+                            <td>
+                                <input type="checkbox" name="deck_id" value="{{ id }}">
+                            </td>
+                            <td>
+                                {{#has_legal_format}}
+                                    {{{ legal_icons }}}
+                                {{/has_legal_format}}
+                            </td>
+                        </tr>
+                    {{/search_results}}
+                </tbody>
+            </table>
+            {{ > archetypedropdown }}
+            <button type="submit">Reassign</button>
+        </form>
+    {{/has_search_results}}
+</section>
